Extract out storage.js
