+++ src/app/user-settings/storage.js
@@ -1,0 +1,47 @@
+/*
+ * Copyright (C) 2018 The "mysteriumnetwork/mysterium-vpn" Authors.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+// @flow
+
+import type { UserSettings } from './user-settings'
+import { readFile, writeFile } from 'fs'
+import { promisify } from 'util'
+
+const readFileAsync = promisify(readFile)
+const writeFileAsync = promisify(writeFile)
+
+async function saveSettings (path: string, settings: UserSettings): Promise<void> {
+  const settingsString = JSON.stringify(settings)
+  await writeFileAsync(path, settingsString)
+}
+
+async function loadSettings (path: string): Promise<UserSettings> {
+  let data = await readFileAsync(path, { encoding: 'utf8' })
+  const parsedSettings = JSON.parse(data)
+
+  if (!validateUserSettings(parsedSettings)) {
+    throw new TypeError('UserSettings loading failed. Parsed Object is not of UserSettings type.')
+  }
+
+  return parsedSettings
+}
+
+function validateUserSettings (settings: Object): boolean {
+  return (typeof settings.showDisconnectNotifications === 'boolean')
+}
+
+export { loadSettings, saveSettings }
